[회원] 사용자 유형 기반 프로필 조회 및 접근 제어 AOP 도입

+++ src/main/java/com/monari/monariback/enrollment/repository/EnrollmentRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface EnrollmentRepository extends JpaRepository<Enrollment, Integer> {
 
-    boolean existsByStudentIdAndLessonId(Long studentId, Integer lessonId);
+    boolean existsByStudentIdAndLessonId(Integer studentId, Integer lessonId);
 }
